+++ eslint.config.js
@@ -4,7 +4,7 @@ import configPrettier from 'eslint-config-prettier';
 import globals from 'globals';
 
 export default tseslint.config(
-  { ignores: ['dist/**', 'coverage/**', 'node_modules/**'] },
+  { ignores: ['dist/**', 'coverage/**', 'node_modules/**', 'examples/**'] },
   js.configs.recommended,
   ...tseslint.configs.recommended,
   {